Moved color sensor code back to launcher
